refactor DenseSet class

+++ src/basic/adt/Hashing.cpp
@@ -17,11 +17,11 @@ namespace basic {
 // Provide a definition and static initializer for the fixed seed. This
 // initializer should always be zero to ensure its value can never appear to be
 // non-zero, even during dynamic initialization.
-size_t hashing::internal::fixed_seed_override = 0;
+uint64_t hashing::internal::fixed_seed_override = 0;
 
 // Implement the function for forced setting of the fixed seed.
 // FIXME: Use atomic operations here so that there is no data race.
-void set_fixed_execution_hash_seed(size_t fixedValue)
+void set_fixed_execution_hash_seed(uint64_t fixedValue)
 {
    hashing::internal::fixed_seed_override = fixedValue;
 }
